+++ sdk/nodejs/migrate/v20171111preview/index.ts
@@ -9,6 +9,7 @@ export * from "./assessment";
 export * from "./getAssessment";
 export * from "./getGroup";
 export * from "./getProject";
+export * from "./getProjectKeys";
 export * from "./group";
 export * from "./project";