+++ sdk/python/pulumi_azure_native/apimanagement/v20161010/__init__.py
@@ -9,5 +9,6 @@ from ._enums import *
 from .api_management_service import *
 from .get_api_management_service import *
 from .get_api_management_service_sso_token import *
+from .get_user_shared_access_token import *
 from ._inputs import *
 from . import outputs