+++ sdk/python/pulumi_azure_native/apimanagement/v20180101/_enums.py
@@ -18,6 +18,7 @@ __all__ = [
     'GroupType',
     'HostnameType',
     'IdentityProviderType',
+    'KeyType',
     'LoggerType',
     'PolicyContentFormat',
     'ProductState',
@@ -140,6 +141,14 @@ class IdentityProviderType(str, Enum):
     TWITTER = "twitter"
     AAD = "aad"
     AAD_B2_C = "aadB2C"
+
+
+class KeyType(str, Enum):
+    """
+    The Key to be used to generate token for user.
+    """
+    PRIMARY = "primary"
+    SECONDARY = "secondary"
 
 
 class LoggerType(str, Enum):