+++ sdk/python/pulumi_azure_native/migrate/v20171111preview/__init__.py
@@ -10,5 +10,6 @@ from .assessment import *
 from .get_assessment import *
 from .get_group import *
 from .get_project import *
+from .get_project_keys import *
 from .group import *
 from .project import *