+++ sdk/python/pulumi_azure_native/operationalinsights/v20200801/__init__.py
@@ -15,6 +15,7 @@ from .get_data_source import *
 from .get_linked_service import *
 from .get_linked_storage_account import *
 from .get_saved_search import *
+from .get_shared_keys import *
 from .get_storage_insight_config import *
 from .get_workspace import *
 from .linked_service import *